+++ js/get-visible-height.ts
@@ -1,13 +1,13 @@
 /**
  * Returns the visible height of an element.
  *
- * @param {HTMLElement} element Element to check the visibility of.
+ * @param {Element} element Element to check the visibility of.
  * @param {number}      offset
  * @param {boolean}     percentage
  *
  * @returns {number}    Visible height of the element in pixels or a percentage of the element's total height.
  */
-export default function getVisibleHeight(element, offset, percentage) {
+export function getVisibleHeight(element: Element, offset: number, percentage: boolean): number {
     const viewportHeight = document.documentElement.clientHeight;
     const rect = element.getBoundingClientRect();
     const rectTopOffset = rect.top - offset;