+++ js/make-jquery-plugin.js
@@ -1,6 +1,0 @@
-export default function makejQueryPlugin($, mostVisible) {
-    // eslint-disable-next-line no-param-reassign, func-names
-    $.fn.mostVisible = function (options) {
-        return this.filter(mostVisible(this.get(), options));
-    };
-}