+++ rollup.config.js
@@ -2,14 +2,32 @@ import terser from '@rollup/plugin-terser';
 import resolve from '@rollup/plugin-node-resolve';
 import babel from '@rollup/plugin-babel';
 
+const extensions = [
+    '.js', '.ts'
+];
+
+const plugins = [
+    resolve({ extensions }),
+    //
+    // // Allow bundling cjs modules. Rollup doesn't understand cjs
+    // commonjs(),
+
+    // Compile TypeScript/JavaScript files
+    babel({
+        extensions,
+        babelHelpers: 'bundled',
+        include: ['js/**/*']
+    })
+];
+
 export default [
     {
-        input: 'js/most-visible.bundle.js',
+        input: 'js/most-visible.bundle.ts',
         output: [
             {
                 file: 'dist/most-visible.js',
                 format: 'iife',
-                name: 'mostVisible',
+                name: 'mostVisible'
             },
             {
                 file: 'dist/most-visible.min.js',
@@ -18,6 +36,20 @@ export default [
                 plugins: [terser()]
             }
         ],
-        plugins: [resolve(), babel({ babelHelpers: 'bundled' })]
+        plugins: plugins
+    },
+    {
+        input: 'js/most-visible.ts',
+        output: [
+            {
+                file: 'dist/most-visible.mjs',
+                format: 'esm'
+            },
+            {
+                file: 'dist/most-visible.cjs',
+                format: 'cjs'
+            }
+        ],
+        plugins: plugins
     }
 ];