+++ CHANGELOG.md
@@ -9,4 +9,4 @@ Version 2 is a major version release with backwards incompatible changes to vers
 * Dropped support for IE11
 * Restructured the plugin away from UMD into two separate files; for the browser (dist/most-visible) and a module for bundlers (js/most-visible).
 * Removed the ability to create an instance of the function with `new mostVisible()`
-* Renamed the `makeJQueryPlugin` function to `makejQueryPlugin`
+* Renamed the `makeJQueryPlugin` function to `makeJqueryPlugin`